% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fflr.R
\name{data_fflr}
\alias{data_fflr}
\title{Fuzzy Fluctuation Logical Relationship}
\usage{
data_fflr(data, N, M)
}
\arguments{
\item{data}{a fuzzified dataset in a dataframe}

\item{N}{order of the Fuzzy Fluctuation Logical Relationship}

\item{M}{Scaling parameter}
}
\value{
A dataframe with only the FFLR
}
\description{
Creates an FFLR matrix.
Populates the matrix, each row is an FFLR, with the last column being the Q(t)_bar value of the FFLR.
Expands the FFLR matrix with (N+1) rows of zeros to match the size of data frame.
Creates a data frame with only the FFLR relationships.
}
