% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fuzzification.R
\name{data_fuzzification}
\alias{data_fuzzification}
\title{Fuzzify the fluctuation time series}
\usage{
data_fuzzification(data)
}
\arguments{
\item{data}{A data frame with columns: Date, AdjClose and Fluctuation}
}
\value{
A data frame containing the fuzzified data
}
\description{
Creates the S(t) vector,
Assigns to each fluctuation its fuzzified value,
Adds S(t) to the data table,
Creates the S(t) bar vector,
Adds S(t) bar to the data table.
}
