% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_analysis.R
\name{error_analysis}
\alias{error_analysis}
\title{Error parameters}
\usage{
error_analysis(data)
}
\arguments{
\item{data}{a data frame containing the actual adjusted close price and the predicted adjusted close price}
}
\description{
Finding the MSE (mean squared error), RMSE (root of the mean squared error), MAE (mean absolute error) and MPE (mean percentage error) for the results of the model
}
