% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer_conversion.R
\name{integer_conversion}
\alias{integer_conversion}
\title{Converting Data Frame Columns to Integer type}
\usage{
integer_conversion(data)
}
\arguments{
\item{data}{a data frame with numeric columns that are not necessarily integers}
}
\value{
The data frame with all but the last column as integers
}
\description{
Changing the data type of all but the last column of a dataframe to integer
}
